allIds, byIds取得処理を関数化
